added all other brands as well
