+++ brands/BOM/tokens/colors.js
@@ -1,0 +1,24 @@
+const COLORS = {
+	background: '#F1F1F2',
+	border: '#D4D4D8',
+	borderDark: '#939299',
+	focus: '#867BCD',
+	heading: '#534891',
+	hero: '#534891',
+	light: '#FBFBFD',
+	muted: '#6E6C7A',
+	neutral: '#595762',
+	primary: '#D73B00',
+	text: '#20024E',
+
+	// reserved
+	success: '#008000',
+	info: '#0074C4',
+	warning: '#C53B00',
+	danger: '#C40000',
+	system: '#ff0',
+};
+
+module.exports = {
+	COLORS,
+};
