+++ brands/BOM/tokens/layout.js
@@ -1,0 +1,10 @@
+module.exports = {
+	LAYOUT: {
+		breakpoints: {
+			sm: 576,
+			md: 768,
+			lg: 992,
+			xl: 1200,
+		},
+	},
+};
