+++ brands/BOM/tokens/type.js
@@ -1,0 +1,33 @@
+module.exports = {
+	TYPE: {
+		brandfonts: [
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/lineto-brown-pro-regular.woff2',
+					woff: '../font/lineto-brown-pro-regular.woff',
+				},
+				weight: 'normal',
+				style: 'normal',
+			},
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/lineto-brown-pro-bold.woff2',
+					woff: '../font/lineto-brown-pro-bold.woff',
+				},
+				weight: 'bold',
+				style: 'normal',
+			},
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/lineto-brown-pro-light.woff2',
+					woff: '../font/lineto-brown-pro-light.woff',
+				},
+				weight: 300,
+				style: 'normal',
+			},
+		],
+	},
+};
