+++ brands/BSA/tokens/colors.js
@@ -1,0 +1,24 @@
+const COLORS = {
+	background: '#F0F1F1',
+	border: '#D3D4D5',
+	borderDark: '#909396',
+	focus: '#C26F9D',
+	heading: '#002F6C',
+	hero: '#002F6C',
+	light: '#F8F8F8',
+	muted: '#6C7074',
+	neutral: '#6C7074',
+	primary: '#D81B2B',
+	text: '#333',
+
+	// reserved
+	success: '#008000',
+	info: '#0074C4',
+	warning: '#C53B00',
+	danger: '#C40000',
+	system: '#ff0',
+};
+
+module.exports = {
+	COLORS,
+};
