+++ brands/BSA/tokens/type.js
@@ -1,0 +1,24 @@
+module.exports = {
+	TYPE: {
+		brandfonts: [
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/Aller_Bd.woff2',
+					woff: '../font/Aller_Bd.woff',
+				},
+				weight: 'normal',
+				style: 'normal',
+			},
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/Aller_Lt.woff2',
+					woff: '../font/Aller_Lt.woff',
+				},
+				weight: 300,
+				style: 'normal',
+			},
+		],
+	},
+};
