+++ brands/BTFG/tokens/colors.js
@@ -1,0 +1,24 @@
+const COLORS = {
+	background: '#F9F9F9',
+	border: '#E0E0E0',
+	borderDark: '#949494',
+	focus: '#CA73A6',
+	heading: '#2574A9',
+	hero: '#2574A9',
+	light: '#FCFCFC',
+	muted: '#666666',
+	neutral: '#637b98',
+	primary: '#9E005D',
+	text: '#333',
+
+	// reserved
+	success: '#008000',
+	info: '#0074C4',
+	warning: '#C53B00',
+	danger: '#C40000',
+	system: '#ff0',
+};
+
+module.exports = {
+	COLORS,
+};
