+++ brands/BTFG/tokens/spacing.js
@@ -1,0 +1,5 @@
+const SPACING = Array.from({ length: 150 }, (_, i) => i * 6);
+
+module.exports = {
+	SPACING,
+};
