+++ brands/STG/index.js
@@ -1,1 +1,4 @@
-export { default } from './tokens/';
+const pkg = require('./package.json');
+const BRAND = pkg.name.replace('@westpac/', '').toUpperCase();
+
+require('../../helpers/transformers/web')(BRAND);
