+++ brands/STG/tokens/colors.js
@@ -1,0 +1,24 @@
+const COLORS = {
+	background: '#F1F1EF',
+	border: '#D6D5D0',
+	borderDark: '#949386',
+	focus: '#9681AF',
+	heading: '#004833',
+	hero: '#78BE20',
+	light: '#F8F8F7',
+	muted: '#757463',
+	neutral: '#757463',
+	primary: '#E30000',
+	text: '#004833',
+
+	// reserved
+	success: '#008000',
+	info: '#0074C4',
+	warning: '#C53B00',
+	danger: '#C40000',
+	system: '#ff0',
+};
+
+module.exports = {
+	COLORS,
+};
