+++ brands/STG/tokens/type.js
@@ -1,0 +1,15 @@
+module.exports = {
+	TYPE: {
+		brandfonts: [
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/dragonbold-bold-webfont.woff2',
+					woff: '../font/dragonbold-bold-webfont.woff',
+				},
+				weight: 'normal',
+				style: 'normal',
+			},
+		],
+	},
+};
