--- brands/STG/tokens/utils.js
@@ -1,6 +1,0 @@
-import Color from 'color';
-
-export const tint = (color, amount) =>
-	Color('white')
-		.mix(Color(color), amount)
-		.hex();
