+++ brands/WBC/tokens/colors.js
@@ -1,0 +1,24 @@
+const COLORS = {
+	background: '#F4F3F0',
+	border: '#D7D2CB',
+	borderDark: '#8C9296',
+	focus: '#9F4585',
+	heading: '#621A4B',
+	hero: '#621A4B',
+	light: '#F9F9F8',
+	muted: '#575F65',
+	neutral: '#2D373E',
+	primary: '#D5002B',
+	text: '#2D373E',
+
+	// reserved
+	success: '#008000',
+	info: '#0074C4',
+	warning: '#C53B00',
+	danger: '#C40000',
+	system: '#ff0',
+};
+
+module.exports = {
+	COLORS,
+};
