+++ brands/WBC/tokens/type.js
@@ -1,0 +1,15 @@
+module.exports = {
+	TYPE: {
+		brandfonts: [
+			{
+				name: 'name',
+				files: {
+					woff2: '../font/7FF5B4E46E46717F5.woff2',
+					woff: '../font/7FF5B4E46E46717F5.woff',
+				},
+				weight: 'normal',
+				style: 'normal',
+			},
+		],
+	},
+};
